refacto : create and implements interfaces IAuthService, IAuthUserService and IEmailService
