+++ src/main/java/com/jeyofdev/yellow_berry/auth/AuthController.java
@@ -13,23 +13,23 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/auth")
 @RequiredArgsConstructor
 public class AuthController {
-    private final AuthService authService;
+    private final AuthServiceImpl authServiceImpl;
 
     @PostMapping("/register")
     public ResponseEntity<RegisterResponse> register(@Valid @RequestBody RegisterRequest request, BindingResult bindingResult) {
-        RegisterResponse registerResponse = authService.register(request, bindingResult);
+        RegisterResponse registerResponse = authServiceImpl.register(request, bindingResult);
         return new ResponseEntity<>(registerResponse, HttpStatus.OK);
     }
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@Valid @RequestBody LoginRequest request, BindingResult bindingResult) {
-        AuthResponse authenticationResponse = authService.login(request, bindingResult);
+        AuthResponse authenticationResponse = authServiceImpl.login(request, bindingResult);
         return new ResponseEntity<>(authenticationResponse, HttpStatus.OK);
     }
 
     @GetMapping("/validate-account")
     public ResponseEntity<MessageResponse> validateAccount(@RequestParam("verificationToken") String verificationToken) {
-        MessageResponse messageResponse = authService.validateAccount(verificationToken);
+        MessageResponse messageResponse = authServiceImpl.validateAccount(verificationToken);
         return new ResponseEntity<>(messageResponse, HttpStatus.OK);
     }
 
@@ -38,7 +38,7 @@ public class AuthController {
             @Valid @RequestBody UpdatePasswordRequest updatePasswordRequest,
             BindingResult bindingResult
     ) {
-        MessageResponse messageResponse = authService.updatePassword(
+        MessageResponse messageResponse = authServiceImpl.updatePassword(
                 updatePasswordRequest.getOldPassword(),
                 updatePasswordRequest.getNewPassword(),
                 bindingResult
@@ -49,7 +49,7 @@ public class AuthController {
 
     @PostMapping("/forgot-password")
     public ResponseEntity<MessageResponse> requestPasswordReset(@RequestParam("email") String email) {
-        MessageResponse requestPasswordResetResponse = authService.requestPasswordReset(email);
+        MessageResponse requestPasswordResetResponse = authServiceImpl.requestPasswordReset(email);
         return new ResponseEntity<>(requestPasswordResetResponse, HttpStatus.OK);
     }
 
@@ -59,7 +59,7 @@ public class AuthController {
             @RequestParam("resetToken") String resetToken,
             @RequestParam("newPassword") String newPassword
     ) {
-        MessageResponse messageResponse = authService.resetPassword(resetToken, newPassword);
+        MessageResponse messageResponse = authServiceImpl.resetPassword(resetToken, newPassword);
         return new ResponseEntity<>(messageResponse, HttpStatus.OK);
     }
 }
